+++ hypervisor/include/arch/x86/cpu.h
@@ -38,6 +38,7 @@
 #ifndef CPU_H
 #define CPU_H
 #include <types.h>
+#include <acrn_common.h>
 
 /* Define CPU stack alignment */
 #define CPU_STACK_ALIGN         16UL
@@ -146,6 +147,41 @@
 
 /* Boot CPU ID */
 #define BOOT_CPU_ID             0U
+
+/* Number of GPRs saved / restored for guest in VCPU structure */
+#define NUM_GPRS                            16U
+#define GUEST_STATE_AREA_SIZE               512
+
+#define	CPU_CONTEXT_OFFSET_RAX			0U
+#define	CPU_CONTEXT_OFFSET_RCX			8U
+#define	CPU_CONTEXT_OFFSET_RDX			16U
+#define	CPU_CONTEXT_OFFSET_RBX			24U
+#define	CPU_CONTEXT_OFFSET_RSP			32U
+#define	CPU_CONTEXT_OFFSET_RBP			40U
+#define	CPU_CONTEXT_OFFSET_RSI			48U
+#define	CPU_CONTEXT_OFFSET_RDI			56U
+#define	CPU_CONTEXT_OFFSET_R8			64U
+#define	CPU_CONTEXT_OFFSET_R9			72U
+#define	CPU_CONTEXT_OFFSET_R10			80U
+#define	CPU_CONTEXT_OFFSET_R11			88U
+#define	CPU_CONTEXT_OFFSET_R12			96U
+#define	CPU_CONTEXT_OFFSET_R13			104U
+#define	CPU_CONTEXT_OFFSET_R14			112U
+#define	CPU_CONTEXT_OFFSET_R15			120U
+#define	CPU_CONTEXT_OFFSET_CR0			128U
+#define	CPU_CONTEXT_OFFSET_CR2			136U
+#define	CPU_CONTEXT_OFFSET_CR4			144U
+#define	CPU_CONTEXT_OFFSET_RIP			152U
+#define	CPU_CONTEXT_OFFSET_RFLAGS		160U
+#define	CPU_CONTEXT_OFFSET_IA32_SPEC_CTRL	168U
+#define	CPU_CONTEXT_OFFSET_IA32_EFER		176U
+#define	CPU_CONTEXT_OFFSET_EXTCTX_START		184U
+#define	CPU_CONTEXT_OFFSET_CR3			184U
+#define	CPU_CONTEXT_OFFSET_IDTR			192U
+#define	CPU_CONTEXT_OFFSET_LDTR			216U
+
+/*sizes of various registers within the VCPU data structure */
+#define VMX_CPU_S_FXSAVE_GUEST_AREA_SIZE    GUEST_STATE_AREA_SIZE
 
 #ifndef ASSEMBLER
 
@@ -268,6 +304,89 @@ enum pcpu_boot_state {
 #define	NEED_SHUTDOWN_VM	(2U)
 void make_pcpu_offline(uint16_t pcpu_id);
 bool need_offline(uint16_t pcpu_id);
+
+struct segment_sel {
+	uint16_t selector;
+	uint64_t base;
+	uint32_t limit;
+	uint32_t attr;
+};
+
+/**
+ * @brief registers info saved for vcpu running context
+ */
+struct run_context {
+/* Contains the guest register set.
+ * NOTE: This must be the first element in the structure, so that the offsets
+ * in vmx_asm.S match
+ */
+	union cpu_regs_t {
+		struct acrn_gp_regs regs;
+		uint64_t longs[NUM_GPRS];
+	} cpu_regs;
+
+	/** The guests CR registers 0, 2, 3 and 4. */
+	uint64_t cr0;
+
+	/* CPU_CONTEXT_OFFSET_CR2 =
+	 * offsetof(struct run_context, cr2) = 136
+	 */
+	uint64_t cr2;
+	uint64_t cr4;
+
+	uint64_t rip;
+	uint64_t rflags;
+
+	/* CPU_CONTEXT_OFFSET_IA32_SPEC_CTRL =
+	 * offsetof(struct run_context, ia32_spec_ctrl) = 168
+	 */
+	uint64_t ia32_spec_ctrl;
+	uint64_t ia32_efer;
+};
+
+/*
+ * extended context does not save/restore during vm exit/entry, it's mainly
+ * used in trusty world switch
+ */
+struct ext_context {
+	uint64_t cr3;
+
+	/* segment registers */
+	struct segment_sel idtr;
+	struct segment_sel ldtr;
+	struct segment_sel gdtr;
+	struct segment_sel tr;
+	struct segment_sel cs;
+	struct segment_sel ss;
+	struct segment_sel ds;
+	struct segment_sel es;
+	struct segment_sel fs;
+	struct segment_sel gs;
+
+	uint64_t ia32_star;
+	uint64_t ia32_lstar;
+	uint64_t ia32_fmask;
+	uint64_t ia32_kernel_gs_base;
+
+	uint64_t ia32_pat;
+	uint32_t ia32_sysenter_cs;
+	uint64_t ia32_sysenter_esp;
+	uint64_t ia32_sysenter_eip;
+	uint64_t ia32_debugctl;
+
+	uint64_t dr7;
+	uint64_t tsc_offset;
+
+	/* The 512 bytes area to save the FPU/MMX/SSE states for the guest */
+	uint64_t
+	fxstore_guest_area[VMX_CPU_S_FXSAVE_GUEST_AREA_SIZE / sizeof(uint64_t)]
+	__aligned(16);
+};
+
+struct cpu_context {
+	struct run_context run_ctx;
+	struct ext_context ext_ctx;
+};
 
 /* Function prototypes */
 void cpu_do_idle(void);
