+++ hypervisor/include/arch/x86/guest/vm.h
@@ -135,7 +135,7 @@ struct acrn_vm {
 	 * so the snapshot only stores the vcpu0's run_context
 	 * of secure world.
 	 */
-	struct cpu_context sworld_snapshot;
+	struct guest_cpu_context sworld_snapshot;
 
 	uint32_t vcpuid_entry_nr, vcpuid_level, vcpuid_xlevel;
 	struct vcpuid_entry vcpuid_entries[MAX_VM_VCPUID_ENTRIES];
